KrushiAI-Disease-Recognition/backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.0
gunicorn==21.2.0
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2
Pillow==10.3.0
joblib==1.3.2
